+++ src/components/DataEntry/DataEntryTable/NewEntry/SenseDialog.tsx
@@ -13,7 +13,7 @@ import { Word } from "api/models";
 import DomainCell from "goals/ReviewEntries/ReviewEntriesComponent/CellComponents/DomainCell";
 import { ReviewEntriesWord } from "goals/ReviewEntries/ReviewEntriesComponent/ReviewEntriesTypes";
 import theme from "types/theme";
-import { firstGlossText } from "types/wordUtilities";
+import { firstGlossText } from "utilities/wordUtilities";
 
 interface SenseDialogProps {
   selectedWord: Word;
