+++ src/components/GoalTimeline/Redux/GoalActions.ts
@@ -13,8 +13,8 @@ import {
 } from "goals/MergeDupGoal/Redux/MergeDupActions";
 import { StoreState } from "types";
 import { StoreStateDispatch } from "types/Redux/actions";
-import { convertEditToGoal } from "types/goalUtilities";
 import { Goal, GoalStatus, GoalType } from "types/goals";
+import { convertEditToGoal } from "utilities/goalUtilities";
 
 // Action Creators
