+++ src/components/GoalTimeline/tests/GoalActions.test.tsx
@@ -16,7 +16,7 @@ import { HandleFlags } from "goals/HandleFlags/HandleFlags";
 import { MergeDups } from "goals/MergeDupGoal/MergeDups";
 import { MergeDupData, MergeStepData } from "goals/MergeDupGoal/MergeDupsTypes";
 import { goalDataMock } from "goals/MergeDupGoal/Redux/tests/MergeDupDataMock";
-import { maxNumSteps } from "types/goalUtilities";
+import { maxNumSteps } from "utilities/goalUtilities";
 import { Goal } from "types/goals";
 import { newUser } from "types/user";
