+++ src/utilities/goalUtilities.ts
