+++ src/utilities/wordUtilities.ts
